requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
openai==1.30.1
flask-swagger-ui==4.11.1
